Sentimenal analysis
This dataset contains tweets posted for various services and products along with the emotion contained in the tweet. It contains three columns, the tweet text, the product/service, and the emotion contained in the tweet. The tweets are used to analyze the product brand and in doing so we are conducting a sentiment analysis. Sentiment analysis is a technique used to understand the positive, negative or neutral sentiments about a particular topic.